requirements.txt
python-dotenv == 1.1.0
streamlit == 1.45.0
openai == 1.77.0